(PSI): move token type related classes to psi package
