+++ src/main/kotlin/org/rust/ide/search/RustWordScanner.kt
@@ -3,7 +3,7 @@ package org.rust.ide.search
 import com.intellij.lang.cacheBuilder.DefaultWordsScanner
 import com.intellij.psi.tree.TokenSet
 import org.rust.lang.core.lexer.RustLexer
-import org.rust.lang.core.lexer.RustTokenElementTypes
+import org.rust.lang.core.psi.RustTokenElementTypes
 
 class RustWordScanner : DefaultWordsScanner(RustLexer(),
     TokenSet.create(RustTokenElementTypes.IDENTIFIER),
