+++ src/main/kotlin/org/rust/lang/core/RustParserDefinition.kt
@@ -13,8 +13,8 @@ import com.intellij.psi.tree.IFileElementType
 import com.intellij.psi.tree.TokenSet
 import org.rust.lang.RustLanguage
 import org.rust.lang.core.lexer.RustLexer
-import org.rust.lang.core.lexer.RustTokenElementTypes
-import org.rust.lang.core.lexer.RustTokenElementTypes.*
+import org.rust.lang.core.psi.RustTokenElementTypes
+import org.rust.lang.core.psi.RustTokenElementTypes.*
 import org.rust.lang.core.parser.RustParser
 import org.rust.lang.core.psi.RustCompositeElementTypes
 import org.rust.lang.core.psi.impl.RustFileImpl
