+++ src/main/kotlin/org/rust/lang/core/psi/RustCompositeElement.kt
@@ -2,7 +2,7 @@ package org.rust.lang.core.psi
 
 import com.intellij.psi.NavigatablePsiElement
 import com.intellij.psi.PsiElement
-import org.rust.lang.core.lexer.RustTokenElementTypes
+import org.rust.lang.core.psi.RustTokenElementTypes
 import org.rust.lang.core.resolve.ref.RustReference
 
 public interface RustCompositeElement   : PsiElement
