+++ src/main/kotlin/org/rust/lang/core/psi/impl/RustNamedElementImpl.kt
@@ -2,7 +2,7 @@ package org.rust.lang.core.psi.impl
 
 import com.intellij.lang.ASTNode
 import com.intellij.psi.PsiElement
-import org.rust.lang.core.lexer.RustTokenElementTypes
+import org.rust.lang.core.psi.RustTokenElementTypes
 import org.rust.lang.core.psi.RustNamedElement
 
 public abstract class RustNamedElementImpl(node: ASTNode)   : RustCompositeElementImpl(node)
