+++ src/main/kotlin/org/rust/lang/core/resolve/ref/RustQualifiedReferenceImpl.kt
@@ -4,7 +4,7 @@ import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReferenceBase
 import org.rust.lang.core.completion.RustCompletionEngine
-import org.rust.lang.core.lexer.RustTokenElementTypes
+import org.rust.lang.core.psi.RustTokenElementTypes
 import org.rust.lang.core.psi.RustNamedElement
 import org.rust.lang.core.psi.RustQualifiedReferenceElement
 import org.rust.lang.core.resolve.RustResolveEngine
